Update FileStorage to Beam's FileSystem
